requirements.txt
python-binance
TA-Lib
numpy~=1.20.2
websocket_client
websocket-client~=0.59.0
backtrader~=1.9.76.123
playsound~=1.2.2
datetime~=4.3

yfinance~=0.1.59
